refactor: move device_name_utils to platform directory and update references.
