+++ tests/core/util/device_name_utils_test.cpp
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#include "core/util/device_name_utils.h"
+#include "core/platform/device_name_utils.h"
 
 #include <gtest/gtest.h>
