+++ xllm/core/distributed_runtime/vlm_master.cpp
@@ -30,6 +30,7 @@ limitations under the License.
 #include "core/common/message.h"
 #include "core/framework/multimodal/mm_data.h"
 #include "core/framework/multimodal/mm_input.h"
+#include "core/platform/device_name_utils.h"
 #include "framework/chat_template/jinja_chat_template.h"
 #include "framework/model/model_args.h"
 #include "framework/request/request.h"
@@ -38,7 +39,6 @@ limitations under the License.
 #include "scheduler/scheduler_factory.h"
 #include "server/xllm_server_registry.h"
 #include "speculative_engine.h"
-#include "util/device_name_utils.h"
 #include "util/scope_guard.h"
 #include "util/timer.h"
 #include "vlm_engine.h"
